cost: refactor instances and lookup function into a seperate package
